[refactoring/daengle-49] 기능별 예외 패키지 구분
