+++ daengle-user-api/src/main/java/ddog/user/application/CareReviewService.java
@@ -11,6 +11,10 @@ import ddog.persistence.mysql.port.ReservationPersist;
 import ddog.persistence.mysql.port.UserPersist;
 import ddog.persistence.mysql.port.VetPersist;
 import ddog.user.application.exception.*;
+import ddog.user.application.exception.estimate.ReservationException;
+import ddog.user.application.exception.estimate.ReservationExceptionType;
+import ddog.user.application.exception.ReviewException;
+import ddog.user.application.exception.ReviewExceptionType;
 import ddog.user.presentation.review.dto.ReviewResp;
 import ddog.user.presentation.review.dto.CareReviewSummaryResp;
 import lombok.RequiredArgsConstructor;
