+++ daengle-user-api/src/main/java/ddog/user/application/EstimateService.java
@@ -7,7 +7,11 @@ import ddog.domain.pet.Pet;
 import ddog.domain.user.User;
 import ddog.domain.vet.Vet;
 import ddog.persistence.mysql.port.*;
-import ddog.user.application.exception.*;
+import ddog.user.application.exception.account.*;
+import ddog.user.application.exception.estimate.CareEstimateException;
+import ddog.user.application.exception.estimate.CareEstimateExceptionType;
+import ddog.user.application.exception.estimate.GroomingEstimateException;
+import ddog.user.application.exception.estimate.GroomingEstimateExceptionType;
 import ddog.user.application.mapper.CareEstimateMapper;
 import ddog.user.application.mapper.GroomingEstimateMapper;
 import ddog.user.presentation.estimate.dto.*;
