+++ daengle-user-api/src/main/java/ddog/user/application/GroomingReviewService.java
@@ -9,6 +9,10 @@ import ddog.domain.user.User;
 import ddog.persistence.mysql.port.GroomerPersist;
 import ddog.persistence.mysql.port.GroomingReviewPersist;
 import ddog.persistence.mysql.port.ReservationPersist;
+import ddog.user.application.exception.estimate.ReservationException;
+import ddog.user.application.exception.estimate.ReservationExceptionType;
+import ddog.user.application.exception.ReviewException;
+import ddog.user.application.exception.ReviewExceptionType;
 import ddog.persistence.mysql.port.UserPersist;
 import ddog.user.application.exception.*;
 import ddog.user.presentation.review.dto.GroomingReviewSummaryResp;
