+++ daengle-user-api/src/main/java/ddog/user/application/ReservationService.java
@@ -2,8 +2,8 @@ package ddog.user.application;
 
 import ddog.domain.payment.Reservation;
 import ddog.persistence.mysql.port.ReservationPersist;
-import ddog.user.application.exception.ReservationException;
-import ddog.user.application.exception.ReservationExceptionType;
+import ddog.user.application.exception.estimate.ReservationException;
+import ddog.user.application.exception.estimate.ReservationExceptionType;
 import ddog.user.presentation.reservation.dto.ReservationSummary;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
