+++ daengle-user-api/src/main/java/ddog/user/application/exception/account/PetException.java
@@ -1,4 +1,4 @@
-package ddog.user.application.exception;
+package ddog.user.application.exception.account;
 
 import ddog.auth.exception.common.CustomRuntimeException;
