+++ daengle-user-api/src/main/java/ddog/user/application/exception/account/PetExceptionType.java
@@ -1,11 +1,12 @@
-package ddog.user.application.exception;
+package ddog.user.application.exception.account;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @AllArgsConstructor
-public enum ReservationExceptionType {
-    RESERVATION_NOT_FOUND(HttpStatus.NOT_FOUND, 404, "예약이 존재하지 않음.");
+public enum PetExceptionType {
+    INVALID_REQUEST_DATA_FORMAT(HttpStatus.BAD_REQUEST, 400, "데이터 형식 오류"),
+    PET_NOT_FOUND(HttpStatus.NOT_FOUND, 2001, "반려동물이 존재하지 않음.");
 
     private final HttpStatus httpStatus;
     private final Integer code;
