+++ daengle-user-api/src/main/java/ddog/user/application/exception/account/UserExceptionType.java
@@ -1,4 +1,4 @@
-package ddog.user.application.exception;
+package ddog.user.application.exception.account;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
