+++ daengle-user-api/src/main/java/ddog/user/application/exception/estimate/ReservationExceptionType.java
@@ -1,12 +1,11 @@
-package ddog.user.application.exception;
+package ddog.user.application.exception.estimate;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @AllArgsConstructor
-public enum GroomingEstimateExceptionType {
-    INVALID_REQUEST_DATA_FORMAT(HttpStatus.BAD_REQUEST, 400, "데이터 형식 오류"),
-    GROOMING_ESTIMATE_NOT_FOUND(HttpStatus.NOT_FOUND, 2001, "미용 견적서가 존재하지 않음.");
+public enum ReservationExceptionType {
+    RESERVATION_NOT_FOUND(HttpStatus.NOT_FOUND, 404, "예약이 존재하지 않음.");
 
     private final HttpStatus httpStatus;
     private final Integer code;
